+++ Docker/initialize.sh
@@ -15,6 +15,12 @@ else
     PGRUNNING=false
 fi
 
+export DEBIAN_FRONTEND=noninteractive
+apt-get clean &> /dev/null
+apt-get -q -y update &> /dev/null
+apt-get -q -y install libsasl2-dev libldap2-dev &> /dev/null
+su -c "source $DA_ACTIVATE && pip install python-ldap &> /dev/null" www-data
+
 # echo "1"
 
 if [ -f /var/run/apache2/apache2.pid ]; then
@@ -422,11 +428,8 @@ fi
 # echo "26"
 
 if [ -n "$PACKAGES" ]; then
-    export DEBIAN_FRONTEND=noninteractive
-    apt-get clean
-    apt-get update
     for PACKAGE in "${PACKAGES[@]}"; do
-	apt-get -q -y install $PACKAGE
+        apt-get -q -y install $PACKAGE &> /dev/null
     done
 fi